Switch ACE compiler build from GCC to VBCC
Unify the toolchain: the runtime library (Makefile-lib) already uses VBCC,
now the compiler (Makefile-ace) does too. This required fixing several
VBCC-compatibility issues in the source:

- Rename CONST typedef to ACECONST (clashes with exec/types.h macro)
- Add #include <string.h> and <stdlib.h> for standard C prototypes
- Add Amiga OS function prototypes (OpenLibrary, CloseLibrary, etc.)
- Add missing ACE internal function prototypes to acedef.h
- Add prototypes for non-standard functions (itoa, ltoa, strupr, fsize)
- Fix ULONG/int args with %ld format strings (cast to long)
- Fix bare returns in int-returning functions (peephole, numericfunction)
- Fix push_struct_var_info return type (char* -> void, unused)
- Fix uninitialized str_item in define_common_or_global_variable
- Fix ULONG IntuitionBase vs NULL pointer comparisons

All 118 tests pass. Build is warning-free except one informational
optimizer-passes notice in statement().

+++ src/ace/c/assign.c
@@ -347,7 +347,7 @@ int    num_derefs,i;
      }
 
      /* offset to array base */
-     sprintf(numbuf,"#%ld",total_offset);
+     sprintf(numbuf,"#%ld",(long)total_offset);
      gen("adda.l",numbuf,"a0");
 
      /* save a0 (will be clobbered by expr) */
@@ -365,7 +365,7 @@ int    num_derefs,i;
         gen("move.l","(sp)+","d0");
 
      /* multiply by element size */
-     sprintf(numbuf,"#%ld",elem_size);
+     sprintf(numbuf,"#%ld",(long)elem_size);
      gen("mulu",numbuf,"d0");
 
      /* restore a0 and add indexed offset */
@@ -387,7 +387,7 @@ int    num_derefs,i;
 
       if (member->type == structptrtype)
       {
-       sprintf(numbuf,"#%ld",post_offset);
+       sprintf(numbuf,"#%ld",(long)post_offset);
        gen("adda.l",numbuf,"a0");
        gen("movea.l","(a0)","a0");
        post_offset = 0;
@@ -428,7 +428,7 @@ int    num_derefs,i;
         { _error(4); return; }
 
      /* push target address onto stack */
-     sprintf(numbuf,"#%ld",post_offset);
+     sprintf(numbuf,"#%ld",(long)post_offset);
      gen("adda.l",numbuf,"a0");
      gen("move.l","a0","-(sp)");
 
@@ -577,7 +577,7 @@ int    num_derefs,i;
       else
       if (member->type == stringtype)  /* string */
       {
-       sprintf(numbuf,"#%ld",total_offset);
+       sprintf(numbuf,"#%ld",(long)total_offset);
        gen("move.l","(sp)+","a1");  /* source */
        gen("adda.l",numbuf,"a0");   /* destination = struct address + offset */
        gen_rt_call("_strcpy");      /* copy source to destination */
