+++ src/ace/c/gfx.c
@@ -812,7 +812,7 @@ BOOL linepatterncalled;
        	gen("move.l",addrbuf,"a0");	/* start address of array */
 
        	/* size of array? */
-	sprintf(numbuf,"#%ld",curr_item->size/2);
+	sprintf(numbuf,"#%ld",(long)(curr_item->size/2));
 	gen("move.l",numbuf,"d0");	/* size of array */
 
         gen("move.l","#0","d1");	/* RESTORE flag */
